Aggiunto commenti + convertito distanza da int a float

+++ app/src/main/java/com/example/runmate/TrainingObject.kt
@@ -1,4 +1,5 @@
 package com.example.runmate
 
-data class TrainingObject(val type: String, val date: String, val startTime: String, val steps: Int, val distance: Int, val calories: Float, val duration: String) {
+// Data class to store trainings
+data class TrainingObject(val type: String, val date: String, val startTime: String, val steps: Int, val distance: Float, val calories: Float, val duration: String) {
 }
